requirements.txt
accelerate==1.1.1
clean-fid==0.1.35
datasets==3.3.0
diffusers==0.30.3
einops==0.8.0
gdown==5.2.0
huggingface-hub==0.24.5
ipykernel==6.29.5
ipython==8.26.0
jupyter==1.0.0
matplotlib==3.9.1
matplotlib-inline==0.1.7
numpy==2.0.1
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.4.0
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.0
seaborn==0.13.2
sentencepiece==0.2.0
statsmodels==0.14.4
torch==2.4.0
torchinfo==1.8.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.3
wandb==0.17.5
-e git+https://github.com/toshas/torch-fidelity.git@master#egg=torch-fidelity